requirements.txt
# 核心依赖
aiohttp>=3.12.0
sounddevice>=0.5.0
numpy>=1.24.0
PyYAML>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# API服务依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0